Fix flaky tests.
There are a few flaky tests here https://cloudup.com/cs68SLifvsj.

* Analytics #trackForm should call submit after a timeout
* cookie #set/#get should set a cookie

These seem to fail on IE mostly. Instead of suppressing these tests, we are going to be more lenient and retry these tests.

Another way we could handle this is by skipping these tests on IE or running this more friendly test only for IE, but I couldn't figure this out in a straightforward way.