setup codecov.io
Adds a new karma configuration that can generate coverage reports that are uploaded to codecov.io.

For some reason, Safari and IE tests fail 100% of the time when code coverage is enabled (likely due to the browserify-istanbul plugin).

Rather than skip those browser or skip coverage entirely, I opted to make a new karma configuration that is exclusively used to report coverage results. This runs as a seperate job on CI and runs only on PhantomJS.