Setup Snyk
Snyk monitors for vulnerable dependencies and notifies us if any were to be found.

This fails the build if any high priority vulnerabilities are found in analytics.js-core.

It runs as a seperate CI job to speed up builds. To reduce the copy paste between the different jobs, I moved the common bits into a defaults section that is shared between the jobs.