LIB-868: Update entity.js to support localStorage
This changes entity.js to dual  write ids to localStorage and the cookies. This applies to userId and groupId.

Reads will check the cookie first, and fallback to the localStorage when a cookie is not available. When cookies are not available during reads, and localStorage is, we write the value to the cookie as well.