feat(ids): check LocalStorage when cookies are not available
This updates entity.js and user.js to dual write ids to localStorage and the cookies. This applies
to userId, groupId and anonymousIds. Reads will check the cookie first, and fallback to the
localStorage when a cookie is not available. When cookies are not available during reads, and
localStorage is, we write the value to the cookie as well.