feat(ids): add `localStorageFallbackDisabled` flag
This adds a flag `localStorageFallbackDisabled` to disable the localStorage fallback code.

`localStorageFallbackDisabled` is falsy by default, hence the fallback code is opt in by default, and requires opt out.

Since the configuration for users and groups are supplied seperately, to use this you need to pass in:

```
analytics.initialize(settings, {
  user: {
    localStorageFallbackDisabled: true,
  },
  group: {
   localStorageFallbackDisabled: true
  }
})
```